+++ packages/transactions/src/index.ts
@@ -46,3 +46,5 @@ export * from './common';
 export * from './signature';
 export * from './structuredDataSignature';
 export * from './postcondition-types';
+
+export * as Pc from './pc';