+++ packages/stacking/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.5.0](https://github.com/hirosystems/stacks.js/compare/v6.4.0...v6.5.0) (2023-03-26)
+
+
+### Bug Fixes
+
+* allow contract principals in stackingclient ([ade6345](https://github.com/hirosystems/stacks.js/commit/ade63453fc5150bc2f849702ec4b814e1a51780c))
+
+
+
+## [6.4.0](https://github.com/hirosystems/stacks.js/compare/v6.3.0...v6.4.0) (2023-03-19)
+
+**Note:** Version bump only for package @stacks/stacking
+
+
+
+
+
+## [6.3.0](https://github.com/hirosystems/stacks.js/compare/v6.2.1...v6.3.0) (2023-03-17)
+
+**Note:** Version bump only for package @stacks/stacking
+
+
+
+
+
+## [6.2.0](https://github.com/hirosystems/stacks.js/compare/v6.1.1...v6.2.0) (2023-02-22)
+
+
+### Bug Fixes
+
+* add missing fee and nonce to all stacking methods ([d81cfd1](https://github.com/hirosystems/stacks.js/commit/d81cfd1b8d1f21e737e6d98bfaa65924b3c871cc))
+
+
+
 ## [6.1.1](https://github.com/hirosystems/stacks.js/compare/v6.1.0...v6.1.1) (2023-01-30)
 
 **Note:** Version bump only for package @stacks/stacking